Remove ccache and Buildroot download directory
The download directory should be automatically set. I bet this was set
by copying Buildroot's .config over at some point, but I'm not sure. It
stores downloaded files in the wrong directory on my system, so I took
it out.

Regarding ccache, I'm in the habit of not committing it enabled since it
causes a lot of files to be added to a users hard drive if they're only
building the system once. It also has a small chance of messing up and I
don't want the cache to be used for any CI builds. (Assuming that at
some point you may build this system on Travis.)